Add files via upload
FlickerFicker 
Originally meant to erase the jpeg meta from an image and change the date the photo was shot. 
In short order, I wanted to forge all of the meta by copying the header from donor image.
What I like about this: The data input and validation.
What I dislike: Requires both jhead and exiv2. Why: I should have used one or the other. I started the script with Jhead and just added to it, instead of rewriting.